changed int to double in coodinates nad added traffic and isOpen variable.

+++ traffic-simulator/src/ija/ija2019/traffic/maps/Coordinate.java
@@ -3,34 +3,34 @@ package ija.ija2019.traffic.maps;
 import java.util.Objects;
 
 public class Coordinate {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Coordinate(int x, int y) {
+    public Coordinate(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public static Coordinate create(int x, int y) {
+    public static Coordinate create(double x, double y) {
         if (x >= 0 && y >= 0) {
             return new Coordinate(x, y);
         }
         return null;
     }
 
-    public int diffX(Coordinate c) {
+    public double diffX(Coordinate c) {
         return this.x - c.x;
     }
 
-    public int diffY(Coordinate c) {
+    public double diffY(Coordinate c) {
         return this.y - c.y;
     }
 
-    public int getX() {
+    public double getX() {
         return this.x;
     }
 
-    public int getY() {
+    public double getY() {
         return this.y;
     }
